% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfun_rpm_23_09_03.R
\name{box_ind}
\alias{box_ind}
\title{box_ind}
\usage{
box_ind(x, newdata)
}
\arguments{
\item{x}{\code{rpms} object}

\item{newdata}{dataframe containing the variables used for the recursive 
partitioning.}
}
\value{
Matrix where each row is a vector of indicators whether
        observation is in box or not.
}
\description{
For each row of data, returns a vector indicators whether
              observation is in that box or not
}
