% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfun_rpm_23_09_03.R
\name{boxes}
\alias{boxes}
\title{boxes}
\usage{
boxes(x)
}
\arguments{
\item{x}{\code{rpms} object}
}
\value{
data.frame including end_node, sample size, splits, and values for each end node
}
\description{
returns end boxes that partition the data
}
