% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_tests.R
\name{domain_est}
\alias{domain_est}
\title{domain_est}
\usage{
domain_est(t1, domdat, weights = rep(1, nrow(domdat)), theta = "mean")
}
\arguments{
\item{t1}{rpms object}

\item{domdat}{data frame rows inside desired domain estimate}

\item{weights}{vector containing the sample weights for each observation}

\item{theta}{string of either "mean" or "total"}
}
\value{
list containing F-statistic and corresponding p-value
}
\description{
Produces estimates of domain mean or total along with the
              variance estimate based on estimated regression tree
}

