% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfun_rpm_18_04_20.R
\name{end_nodes}
\alias{end_nodes}
\alias{rpms::end_nodes}
\title{end_nodes}
\usage{
end_nodes(t1)
}
\arguments{
\item{t1}{\code{rpms} object}
}
\value{
vector of lables for each end-node.
}
\description{
Get vector of end-node labels
}
\examples{
{
# model mean of retirement account value for households with reported 
# retirment account values > 0 using a binary tree while accounting for 
# clusterd data and sample weights.

s1<- which(CE$IRAX > 0)
r1 <-rpms(IRAX~EDUCA+AGE+BLS_URBN, data = CE[s1,],  weights=~FINLWT21, clusters=~CID) 
 
 end_nodes(r1)
}

}
