% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfun_rpm_23_09_03.R
\name{grow_rpms}
\alias{grow_rpms}
\title{grow_rpms}
\usage{
grow_rpms(
  x,
  node,
  data,
  weights = ~1,
  strata = ~1,
  clusters = ~1,
  pval = NA,
  bin_size = NA
)
}
\arguments{
\item{x}{\code{rpms} object}

\item{node}{node from which to grow tree further}

\item{data}{data.frame that includes variables used in rp_equ, e_equ, 
and design information}

\item{weights}{formula or vector of sample weights for each observation}

\item{strata}{formula or vector of strata labels}

\item{clusters}{formula or vector of cluster labels}

\item{pval}{numeric p-value used to reject null hypothesis in permutation test}

\item{bin_size}{numeric minimum number of observations in each node}
}
\value{
rpms tree expanded from node.
}
\description{
grow an rpms tree from a given node
}
