% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfun_rpm_23_09_03.R
\name{length.rpms}
\alias{length.rpms}
\alias{length}
\title{length.rpms}
\usage{
\method{length}{rpms}(x)
}
\arguments{
\item{x}{rpms tree model object}
}
\value{
number of end-nodes of the tree
}
\description{
length.rpms
}
