% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfun_rpm_23_09_03.R
\name{linearize}
\alias{linearize}
\title{linearize}
\usage{
linearize(x, data, weights = ~1, strata = ~1, clusters = ~1, type = "part")
}
\arguments{
\item{x}{\code{rpms} object}

\item{data}{data.frame}

\item{weights}{formula or vector of sample weights for each observation}

\item{strata}{formula or vector of strata labels}

\item{clusters}{formula or vector of cluster labels}

\item{type}{is on of "part" or "lin"}
}
\value{
data.frame including splits and estimates for the coefficient and their standard errors
}
\description{
returns a linerized version of the splits. The coefficients represent 
              the effect that each split has on the mean
}
