% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfun_rpm_23_09_03.R
\name{plot.rpms}
\alias{plot.rpms}
\alias{plot}
\title{plot.rpms}
\usage{
\method{plot}{rpms}(x, ...)
}
\arguments{
\item{x}{\code{rpms} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
plot method for class \code{rpms}
}
\details{
Function to plot rpms tree structure using frame information
}
