% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfun_rpm_0_1.R
\name{predict.rpms}
\alias{predict}
\alias{predict.rpms}
\title{predict.rpms}
\usage{
\method{predict}{rpms}(object, newdata, ...)
}
\arguments{
\item{object}{Object inheriting from  \code{rpms}}

\item{newdata}{data frame with variables to use for predicting new values.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
vector of predicticed values for each row of newdata
}
\description{
Predicted values based on \code{rpms} object
}
\examples{
{
# get rpms model of mean retirement contribution by several factors
r1 <- rpms(FINDRETX~EDUC_REF+AGE_REF+BLS_URBN+REGION, data = CE)

# first 10 predicted means
predict(r1, CE[1:10, ])

}

}

