% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_21_08_31.R
\name{predict.rpms_forest}
\alias{predict.rpms_forest}
\title{predict.rpms_forest}
\usage{
\method{predict}{rpms_forest}(object, newdata, ...)
}
\arguments{
\item{object}{Object inheriting from  \code{rpms_forest}}

\item{newdata}{data frame with variables to use for predicting new values.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
vector of predicticed values for each row of newdata
}
\description{
Gets predicted values given new data based on \code{rpms_forest} model.
}
