% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_21_08_31.R
\name{predict.rpms_proj}
\alias{predict.rpms_proj}
\title{predict.rpms_proj}
\usage{
\method{predict}{rpms_proj}(object, newdata, ...)
}
\arguments{
\item{object}{Object inheriting from  \code{rpms_zinf}}

\item{newdata}{data frame with variables to use for predicting new values.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
vector of predicticed values for each row of newdata
}
\description{
Predicted values based on \code{rpms_zinf} model
}
