% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfun_rpm_23_09_03.R
\name{print.rpms}
\alias{print.rpms}
\title{print.rpms}
\usage{
\method{print}{rpms}(x, ...)
}
\arguments{
\item{x}{\code{rpms} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print method for class \code{rpms}
}
