% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_21_08_31.R
\name{print.rpms_forest}
\alias{print.rpms_forest}
\title{print.rpms_forest}
\usage{
\method{print}{rpms_forest}(x, ...)
}
\arguments{
\item{x}{Object inheriting from  \code{rpms_forest}}

\item{...}{further arguments passed to or from other methods.}
}
\value{
vector of predicticed values for each row of newdata
}
\description{
Prints information for a given \code{rpms_forest} model.
}
