% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_21_08_31.R
\name{print.rpms_zinf}
\alias{print.rpms_zinf}
\title{print.rpms_zinf}
\usage{
\method{print}{rpms_zinf}(x, ...)
}
\arguments{
\item{x}{\code{rpms_zinf} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print method for class \code{rpms_zinf}
}
