% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfun_rpm_23_09_03.R
\name{prune_rpms}
\alias{prune_rpms}
\title{prune_rpms}
\usage{
prune_rpms(x, node)
}
\arguments{
\item{x}{\code{rpms} object}

\item{node}{number of node to prune to.}
}
\value{
subtree ending clipping off any splits after given node.
}
\description{
prune rpms tree to given node
}
