% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_21_08_31.R
\name{r2stat}
\alias{r2stat}
\title{r2}
\usage{
r2stat(t1, data, adjusted = TRUE)
}
\arguments{
\item{t1}{Object inheriting from  \code{rpms} \code{rpms_forest}
\code{rpms_boost} or \code{rpms_zinf}}

\item{data}{data frame with variables used to estimate model}

\item{adjusted}{TRUE/FALSE whether to compute adjusted R^2}
}
\value{
R^2 statistic computed using the model and provided data
}
\description{
Returns the estimated R^2 statistic for determining
              the fit of the given model to the data
}
