% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_21_08_31.R
\name{rpms_forest}
\alias{rpms_forest}
\title{rpms_forest}
\usage{
rpms_forest(
  rp_equ,
  data,
  weights = ~1,
  strata = ~1,
  clusters = ~1,
  e_fn = "survLm",
  l_fn = NULL,
  bin_size = 5,
  f_size = 500,
  cores = 1
)
}
\arguments{
\item{rp_equ}{formula containing all variables for partitioning}

\item{data}{data.frame that includes variables used in rp_equ, e_equ, 
and design information}

\item{weights}{formula or vector of sample weights for each observation}

\item{strata}{formula or vector of strata labels}

\item{clusters}{formula or vector of cluster labels}

\item{e_fn}{string name of function to use for modeling 
(only "survLm" is operational)}

\item{l_fn}{loss function (ignored)}

\item{bin_size}{numeric minimum number of observations in each node}

\item{f_size}{integer specifying the number of trees in the forest}

\item{cores}{integer number of cores to use in parallel if > 1 
(doesn't work with Windows operating systems)}
}
\value{
object of class "rpms"
}
\description{
produces a random forest using rpms to create the individual
             trees.
}
