% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_21_08_31.R
\name{rpms_proj}
\alias{rpms_proj}
\title{rpms_proj}
\usage{
rpms_proj(object, newdata, weights = ~1, strata = ~1, clusters = ~1)
}
\arguments{
\item{object}{Object inheriting from  \code{rpms}}

\item{newdata}{data frame with variables used to estimate model}

\item{weights}{formula or vector of sample weights for each observation}

\item{strata}{formula or vector of strata labels}

\item{clusters}{formula or vector of cluster labels}
}
\value{
survLm_fit object
}
\description{
Returns a survLm_fit object with coeficients projecting 
              new data onto splits from the given rpms model.
}
