% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_21_08_31.R
\name{rpms_zinf}
\alias{rpms_zinf}
\title{rpms_zinf}
\usage{
rpms_zinf(
  rp_equ,
  data,
  weights = ~1,
  strata = ~1,
  clusters = ~1,
  e_equ = ~1,
  e_fn = "survLm",
  l_fn = NULL,
  bin_size = NULL,
  gridpts = 3,
  perm_reps = 1000L,
  pval = 0.05
)
}
\arguments{
\item{rp_equ}{formula containing all variables for partitioning}

\item{data}{data.frame that includes variables used in rp_equ, e_equ, 
and design information}

\item{weights}{formula or vector of sample weights for each observation}

\item{strata}{formula or vector of strata labels}

\item{clusters}{formula or vector of cluster labels}

\item{e_equ}{formula for modeling data in each node}

\item{e_fn}{string name of function to use for modeling 
(only "survLm" is operational)}

\item{l_fn}{loss function (does nothing yet)}

\item{bin_size}{numeric minimum number of observations in each node}

\item{gridpts}{integer number of middle points to do in search}

\item{perm_reps}{integer specifying the number of thousands of permuation 
replications to use to estimate p-value}

\item{pval}{numeric p-value used to reject null hypothesis in permutation 
test}
}
\value{
object of class "rpms"
}
\description{
main function producing a regression tree using variables
 from rp_equ to partition the data and fit the model e_equ on each node.
 Currently only uses data with complete cases.
}
