% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_0_1_0.R
\name{split}
\alias{split}
\title{split}
\usage{
split(node, data, e_equ, e_fn = survLm, l_fn, X, weights, strata, clusters,
  bin_size, perm_reps, pval)
}
\arguments{
\item{node}{is an integer label of the node}

\item{data}{this is the data}

\item{e_equ}{is the estimating equation used to model the data in each node}

\item{e_fn}{is the function that should be used to fit the model in each
node.  The function should}

\item{l_fn}{function taking a numeric vector and returning the estimated loss}

\item{X}{vector of string names for variables to consider for splitting}

\item{weights}{the sample weights for each observation}

\item{strata}{lable of the strata containing the observation}

\item{clusters}{lable of the clusters containing the observation}

\item{bin_size}{integer specifying the minimum number of observations each
node must contain}

\item{perm_reps}{integer specifying the number of permuations}

\item{pval}{numeric p-value used to reject null hypothesis in permutation 
test}
}
\value{
list containing split information
}
\description{
internal split function for RPMS
}
