% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfun_rpm_23_09_03.R
\name{survLm}
\alias{survLm}
\title{survLm}
\usage{
survLm(
  e_equ,
  data,
  weights = rep(1, nrow(data)),
  strata = rep(1L, nrow(data)),
  clusters = (1L:nrow(data))
)
}
\arguments{
\item{e_equ}{formula representing the equation to fit}

\item{data}{data.frame}

\item{weights}{formula or vector of sample weights for each observation}

\item{strata}{formula or vector of strata labels}

\item{clusters}{formula or vector of cluster labels}
}
\value{
list containing coefficients, covariance matrix and the residuals
}
\description{
wrapper function for the C++ function survLm_model
}
\keyword{internal}
