% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{survLm_model}
\alias{survLm_model}
\title{Fit a linear model using data collected from a complex sample}
\usage{
survLm_model(y, X, weights, strata, clusters)
}
\arguments{
\item{y}{A vector of values}

\item{X}{The design matrix of the linear model}

\item{weights}{A vector of sample weights for each observation}

\item{strata}{A vector of strata labels}

\item{clusters}{A vector of cluster labels}
}
\value{
list containing coefficients, covariance matrix and the residuals
}
\description{
Fit a linear model using data collected from a complex sample
}
\keyword{internal}
