% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_tests.R
\name{ttest}
\alias{ttest}
\title{ttest}
\usage{
ttest(t1, data, dom, weights)
}
\arguments{
\item{t1}{rpms object}

\item{data}{data frame}

\item{dom}{vector of indices of the data frame that define the domain}

\item{weights}{vector of sample weights for each observation}
}
\value{
list containing the overall effect of the weights on the domain 
        estimate, the t-statistic, and the corresponding p-value
}
\description{
Conducts a t-test for the signifcance of the weights on 
              on the domain estimate
}

