% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_0_1_0.R
\name{var_select}
\alias{var_select}
\title{var_select}
\usage{
var_select(e_equ, data, e_fn = survLm_fit, weights = rep(1, nrow(data)),
  strata = rep(1, nrow(data)), clusters = (1:nrow(data)), des_ind, X,
  perm_reps, pval)
}
\arguments{
\item{e_equ}{formula object of the estimating equation}

\item{data}{dataframe containing partitioning and estimating variables
as well as any used sample design variables}

\item{e_fn}{error function}

\item{weights}{the sample weights for each observation}

\item{strata}{lable of the strata containing the observation}

\item{clusters}{lable of the clusters containing the observation}

\item{X, }{string vector containing the names of pontential variables}

\item{perm_reps}{integer specifying the number of permuations}

\item{pval}{numeric p-value used to reject null hypothesis in permutation 
test}
}
\value{
vector of variable names with lowest estimated p-value
}
\description{
selects variables likely to result in the best split
}
\keyword{internal}
