% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_tests.R
\name{Wtest}
\alias{Wtest}
\alias{Wald}
\alias{test}
\title{Wtest (Wald Test for sample weights)}
\usage{
Wtest(t1, data, weights = rep(1, nrow(data)), strata = rep(1L, nrow(data)),
  clusters = (1L:nrow(data)))
}
\arguments{
\item{t1}{rpms object}

\item{data}{data.frame containing the data}

\item{weights}{vector of sample weights for each observation}

\item{strata}{vector of strata labels}

\item{clusters}{vector of cluster labels}
}
\value{
list containing F-statistic and corresponding p-value
}
\description{
Conducts a Wald's test for the significance of the weights on the splits
}
