% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfun_rpm_0_1.R
\name{end_nodes}
\alias{end_nodes}
\alias{rpms::end_nodes}
\title{end_nodes}
\usage{
end_nodes(t1)
}
\arguments{
\item{t1}{\code{rpms} object}
}
\value{
vector of lables of end-nodes.
}
\description{
Get vector end-node labels
}
