% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfun_rpm_23_09_03.R
\name{end_nodes}
\alias{end_nodes}
\title{end_nodes}
\usage{
end_nodes(object, newdata = NULL)
}
\arguments{
\item{object}{\code{rpms} object}

\item{newdata}{data.frame}
}
\value{
vector of end_node labels
}
\description{
Either a vector of end-node labels for each opbservation in newdata or 
              a vector of the endnodes in the tree model if newdata is not provided.
}
\examples{
{
# model mean of retirement account value for households with reported 
# retirment account values > 0 using a binary tree while accounting for 
# clusterd data and sample weights.

s1<- which(CE$IRAX > 0)
r1 <-rpms(IRAX~EDUCA+AGE+BLS_URBN, data = CE[s1,],  weights=~FINLWT21, clusters=~CID) 
 
 end_nodes(r1)
}

}
